requirements.txt
python-owasp-zap-v2.4==0.1.0
pandas>=1.3.0
scikit-learn>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.0.0
tqdm>=4.62.0